backend/requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
build==1.4.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.5.0
Deprecated==1.3.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
fastapi==0.128.0
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-cloud-vision==3.12.0
google-genai==1.60.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
instructor==1.14.4
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.6.2
kubernetes==35.0.0
langchain==1.2.7
langgraph
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-ollama==1.0.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
lxml==6.0.2
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
msoffcrypto-tool==6.0.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
neo4j==6.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
nodeenv==1.10.0
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
olefile==0.47
ollama==0.6.1
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
openai==2.15.0
opencv-python==4.13.0.90
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pandas==3.0.0
pdf2image==1.17.0
pdfminer.six==20260107
pi_heif==1.2.0
pikepdf==10.2.0
pillow==12.1.0
pillow_heif==1.2.0
platformdirs==4.5.1
portalocker==3.2.0
posthog==5.4.0
pre_commit==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.7
pypandoc==1.16.2
pyparsing==3.3.2
pypdf==6.6.0
pypdfium2==5.3.0
PyPika==0.50.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
ragas==0.4.3
RapidFuzz==3.14.3
redis==6.4.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.10.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.24
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==4.57.6
triton==3.6.0
ty==0.0.14
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unstructured==0.18.27
unstructured-client==0.42.8
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.7
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
virtualenv==20.36.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.3.5
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
llama-parse>=0.5.0
llama-index-core>=0.12.0
